Added ModelCollector object
