+++ Assets/Dodgeball/Scripts/CoPlayModels.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 27699da2c5d5a4c04a5ce10acaa862de
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
